signals: Don't collect siginfo_t-s on stack
We've moved signinfos on core entry, thus the bits with
siginfo-s themselves cannot sit on stack any longer.
Otherwise we would overwritem them with next batch and
will feed stack pointer to the caller, thus causing a
data and garbage on the stack to be written into image
instead of siginfo data.

Signed-off-by: Pavel Emelyanov <xemul@parallels.com>