criu: Version 1.3.1
We have a couple of problems missed in 1.3, so here's
the -stable .1 for that release.

First of all, we nail down the way CRIU decides whether
restore root task as child or as sibling with the
explicit API switch.

And there are two nasty issues in how CRIU dumps mountpoints.

Signed-off-by: Pavel Emelyanov <xemul@parallels.com>