tun: rename FD_TYPES__TUN to FD_TYPES__TUNF
We have two global enums with the member TUN. The first one is in fdinfo.proto and the
second one is in netdev.proto. protoc complains about having TUN in 2 global enums because
of the enum rules for C++(even though we're compiling .proto files for python).
So lets rename TUN into TUNF to avoid that conflict.

Signed-off-by: Ruslan Kuprieiev <kupruser@gmail.com>
Signed-off-by: Pavel Emelyanov <xemul@parallels.com>